+++ brickrail-gui/layout/track/layout_sensor.gd
@@ -2,12 +2,18 @@ extends Reference
 class_name LayoutSensor
 
 var marker_color
+var highlight = 0
 
 signal marker_color_changed()
+signal highlight_changed()
 signal triggered(train)
 
 func _init(p_marker_color=null):
 	marker_color = p_marker_color
+
+func increment_highlight(val):
+	highlight += val
+	emit_signal("highlight_changed")
 
 func get_color():
 	if marker_color == null: